+++ src/en/reanime/build.gradle
@@ -9,5 +9,5 @@ apply plugin: "kei.plugins.extension.legacy"
 
 dependencies {
     implementation(project(":lib:playlistutils"))
-    implementation "org.nanohttpd:nanohttpd:2.3.1"
+    implementation("org.nanohttpd:nanohttpd:2.3.1")
 }
